feat: add maintenance mode middleware
Add a new middleware to handle maintenance mode. The middleware checks if the
`isUnderMaintenance` flag is set in the `@vercel/edge-config` package. If the
flag is set, the middleware rewrites the request URL to the `_maintenance` page.

Also, add the `@vercel/edge-config` package as a dependency in the `package.json`
file.

Finally, add a `_maintenance` page to display a "Coming soon™" message.