chore: update Node.js version in .nvmrc and package.json feat: update dependencies in package.json
The Node.js version in the .nvmrc file and the engines section of the package.json file were updated from 18.18.1 to 20.10.0. This was done to ensure compatibility with the latest version of Node.js.

In the package.json file, several dependencies were updated to their latest versions:
- "@sentry/nextjs" was updated from 7.77.0 to 7.84.0
- "i18next" was updated from 23.5.1 to 23.7.7
- "next" was updated from 13.5.4 to 14.0.3
- "next-i18next" was updated from 14.0.3 to 15.0.0
- "react-i18next" was updated from 13.3.1 to 13.5.0

These updates were done to ensure that the project is using the latest versions of its dependencies, which may include bug fixes, performance improvements, and new features.